+++ sentry-graphql/src/main/java/io/sentry/graphql/SentryDataFetcherExceptionHandler.java
@@ -5,11 +5,10 @@ import static io.sentry.TypeCheckHint.GRAPHQL_HANDLER_PARAMETERS;
 import graphql.execution.DataFetcherExceptionHandler;
 import graphql.execution.DataFetcherExceptionHandlerParameters;
 import graphql.execution.DataFetcherExceptionHandlerResult;
+import io.sentry.Hint;
 import io.sentry.HubAdapter;
 import io.sentry.IHub;
 import io.sentry.util.Objects;
-import java.util.HashMap;
-import java.util.Map;
 import org.jetbrains.annotations.NotNull;
 
 /**
@@ -34,10 +33,10 @@ public final class SentryDataFetcherExceptionHandler implements DataFetcherExcep
   @SuppressWarnings("deprecation")
   public DataFetcherExceptionHandlerResult onException(
       final @NotNull DataFetcherExceptionHandlerParameters handlerParameters) {
-    final Map<String, Object> hintMap = new HashMap<>();
-    hintMap.put(GRAPHQL_HANDLER_PARAMETERS, handlerParameters);
+    final Hint hint = new Hint();
+    hint.set(GRAPHQL_HANDLER_PARAMETERS, handlerParameters);
 
-    hub.captureException(handlerParameters.getException(), hintMap);
+    hub.captureException(handlerParameters.getException(), hint);
     return delegate.onException(handlerParameters);
   }
 }