+++ sentry-samples/sentry-samples-spring-boot/src/main/java/io/sentry/samples/spring/boot/CustomEventProcessor.java
@@ -1,11 +1,10 @@
 package io.sentry.samples.spring.boot;
 
 import io.sentry.EventProcessor;
+import io.sentry.Hint;
 import io.sentry.SentryEvent;
 import io.sentry.protocol.SentryRuntime;
-import java.util.Map;
 import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
 import org.springframework.boot.SpringBootVersion;
 import org.springframework.stereotype.Component;
 
@@ -26,8 +25,7 @@ public class CustomEventProcessor implements EventProcessor {
   }
 
   @Override
-  public @NotNull SentryEvent process(
-      @NotNull SentryEvent event, @Nullable Map<String, Object> hint) {
+  public @NotNull SentryEvent process(@NotNull SentryEvent event, @NotNull Hint hint) {
     final SentryRuntime runtime = new SentryRuntime();
     runtime.setVersion(springBootVersion);
     runtime.setName("Spring Boot");