+++ sentry-servlet-jakarta/src/test/kotlin/io/sentry/servlet/jakarta/SentryRequestHttpServletRequestProcessorTest.kt
@@ -3,6 +3,7 @@ package io.sentry.servlet.jakarta
 import com.nhaarman.mockitokotlin2.any
 import com.nhaarman.mockitokotlin2.mock
 import com.nhaarman.mockitokotlin2.whenever
+import io.sentry.Hint
 import io.sentry.SentryEvent
 import io.sentry.SentryOptions
 import jakarta.servlet.http.HttpServletRequest
@@ -29,7 +30,7 @@ class SentryRequestHttpServletRequestProcessorTest {
         val eventProcessor = SentryRequestHttpServletRequestProcessor(request)
         val event = SentryEvent()
 
-        eventProcessor.process(event, null)
+        eventProcessor.process(event, Hint())
 
         assertNotNull(event.request)
         val eventRequest = event.request!!
@@ -56,7 +57,7 @@ class SentryRequestHttpServletRequestProcessorTest {
         val eventProcessor = SentryRequestHttpServletRequestProcessor(request)
         val event = SentryEvent()
 
-        eventProcessor.process(event, null)
+        eventProcessor.process(event, Hint())
 
         assertNotNull(event.request) {
             assertEquals(
@@ -81,7 +82,7 @@ class SentryRequestHttpServletRequestProcessorTest {
         val eventProcessor = SentryRequestHttpServletRequestProcessor(request)
         val event = SentryEvent()
 
-        eventProcessor.process(event, null)
+        eventProcessor.process(event, Hint())
 
         assertNotNull(event.request) {
             assertNull(it.cookies)
@@ -105,7 +106,7 @@ class SentryRequestHttpServletRequestProcessorTest {
         val eventProcessor = SentryRequestHttpServletRequestProcessor(request)
         val event = SentryEvent()
 
-        eventProcessor.process(event, null)
+        eventProcessor.process(event, Hint())
 
         assertNotNull(event.request) { req ->
             assertNotNull(req.headers) {