+++ sentry-spring/src/main/java/io/sentry/spring/tracing/SentrySpanClientHttpRequestInterceptor.java
@@ -6,6 +6,7 @@ import static io.sentry.TypeCheckHint.SPRING_REQUEST_INTERCEPTOR_RESPONSE;
 
 import com.jakewharton.nopen.annotation.Open;
 import io.sentry.Breadcrumb;
+import io.sentry.Hint;
 import io.sentry.IHub;
 import io.sentry.ISpan;
 import io.sentry.SentryTraceHeader;
@@ -13,8 +14,6 @@ import io.sentry.SpanStatus;
 import io.sentry.TracingOrigins;
 import io.sentry.util.Objects;
 import java.io.IOException;
-import java.util.HashMap;
-import java.util.Map;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import org.springframework.http.HttpRequest;
@@ -81,13 +80,13 @@ public class SentrySpanClientHttpRequestInterceptor implements ClientHttpRequest
         Breadcrumb.http(request.getURI().toString(), request.getMethodValue(), responseStatusCode);
     breadcrumb.setData("request_body_size", body.length);
 
-    final Map<String, Object> hintMap = new HashMap<>();
-    hintMap.put(SPRING_REQUEST_INTERCEPTOR_REQUEST, request);
-    hintMap.put(SPRING_REQUEST_INTERCEPTOR_REQUEST_BODY, body);
+    final Hint hint = new Hint();
+    hint.set(SPRING_REQUEST_INTERCEPTOR_REQUEST, request);
+    hint.set(SPRING_REQUEST_INTERCEPTOR_REQUEST_BODY, body);
     if (response != null) {
-      hintMap.put(SPRING_REQUEST_INTERCEPTOR_RESPONSE, response);
+      hint.set(SPRING_REQUEST_INTERCEPTOR_RESPONSE, response);
     }
 
-    hub.addBreadcrumb(breadcrumb, hintMap);
+    hub.addBreadcrumb(breadcrumb, hint);
   }
 }