+++ sentry/src/main/java/io/sentry/Hub.java
@@ -1,7 +1,5 @@
 package io.sentry;
 
-import static io.sentry.TypeCheckHint.SENTRY_TYPE_CHECK_HINT;
-
 import io.sentry.Stack.StackItem;
 import io.sentry.clientreport.DiscardReason;
 import io.sentry.hints.SessionEndHint;
@@ -10,12 +8,12 @@ import io.sentry.protocol.SentryId;
 import io.sentry.protocol.SentryTransaction;
 import io.sentry.protocol.User;
 import io.sentry.util.ExceptionUtils;
+import io.sentry.util.HintUtils;
 import io.sentry.util.Objects;
 import io.sentry.util.Pair;
 import java.io.Closeable;
 import java.util.Collections;
 import java.util.Date;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.WeakHashMap;
@@ -77,7 +75,7 @@ public final class Hub implements IHub {
 
   @Override
   public @NotNull SentryId captureEvent(
-      final @NotNull SentryEvent event, final @Nullable Map<String, Object> hint) {
+      final @NotNull SentryEvent event, final @Nullable Hint hint) {
     SentryId sentryId = SentryId.EMPTY_ID;
     if (!isEnabled()) {
       options
@@ -129,7 +127,7 @@ public final class Hub implements IHub {
   @ApiStatus.Internal
   @Override
   public @NotNull SentryId captureEnvelope(
-      final @NotNull SentryEnvelope envelope, final @Nullable Map<String, Object> hint) {
+      final @NotNull SentryEnvelope envelope, final @Nullable Hint hint) {
     Objects.requireNonNull(envelope, "SentryEnvelope is required.");
 
     SentryId sentryId = SentryId.EMPTY_ID;
@@ -155,7 +153,7 @@ public final class Hub implements IHub {
 
   @Override
   public @NotNull SentryId captureException(
-      final @NotNull Throwable throwable, final @Nullable Map<String, Object> hint) {
+      final @NotNull Throwable throwable, final @Nullable Hint hint) {
     SentryId sentryId = SentryId.EMPTY_ID;
     if (!isEnabled()) {
       options
@@ -237,16 +235,14 @@ public final class Hub implements IHub {
         // single envelope
         // Or create the envelope here with both items and call `captureEnvelope`
         if (pair.getPrevious() != null) {
-          final Map<String, Object> hintMap = new HashMap<>();
-          hintMap.put(SENTRY_TYPE_CHECK_HINT, new SessionEndHint());
+          final Hint hint = HintUtils.createWithTypeCheckHint(new SessionEndHint());
 
-          item.getClient().captureSession(pair.getPrevious(), hintMap);
+          item.getClient().captureSession(pair.getPrevious(), hint);
         }
 
-        final Map<String, Object> hintMap = new HashMap<>();
-        hintMap.put(SENTRY_TYPE_CHECK_HINT, new SessionStartHint());
+        final Hint hint = HintUtils.createWithTypeCheckHint(new SessionStartHint());
 
-        item.getClient().captureSession(pair.getCurrent(), hintMap);
+        item.getClient().captureSession(pair.getCurrent(), hint);
       } else {
         options.getLogger().log(SentryLevel.WARNING, "Session could not be started.");
       }
@@ -263,10 +259,9 @@ public final class Hub implements IHub {
       final StackItem item = this.stack.peek();
       final Session previousSession = item.getScope().endSession();
       if (previousSession != null) {
-        final Map<String, Object> hintMap = new HashMap<>();
-        hintMap.put(SENTRY_TYPE_CHECK_HINT, new SessionEndHint());
+        final Hint hint = HintUtils.createWithTypeCheckHint(new SessionEndHint());
 
-        item.getClient().captureSession(previousSession, hintMap);
+        item.getClient().captureSession(previousSession, hint);
       }
     }
   }
@@ -298,8 +293,7 @@ public final class Hub implements IHub {
   }
 
   @Override
-  public void addBreadcrumb(
-      final @NotNull Breadcrumb breadcrumb, final @Nullable Map<String, Object> hint) {
+  public void addBreadcrumb(final @NotNull Breadcrumb breadcrumb, final @Nullable Hint hint) {
     if (!isEnabled()) {
       options
           .getLogger()
@@ -552,7 +546,7 @@ public final class Hub implements IHub {
   public @NotNull SentryId captureTransaction(
       final @NotNull SentryTransaction transaction,
       final @Nullable TraceState traceState,
-      final @Nullable Map<String, Object> hint,
+      final @Nullable Hint hint,
       final @Nullable ProfilingTraceData profilingTraceData) {
     Objects.requireNonNull(transaction, "transaction is required");