+++ sentry/src/test/java/io/sentry/DuplicateEventDetectionEventProcessorTest.kt
@@ -25,9 +25,12 @@ class DuplicateEventDetectionEventProcessorTest {
     @Test
     fun `does not drop event if no previous event with same exception was processed`() {
         val processor = fixture.getSut()
-        processor.process(SentryEvent(), null)
+        processor.process(SentryEvent(), Hint())
 
-        val result = processor.process(SentryEvent(RuntimeException()), null)
+        val result = processor.process(
+            SentryEvent(RuntimeException()),
+            Hint()
+        )
 
         assertNotNull(result)
     }
@@ -36,9 +39,9 @@ class DuplicateEventDetectionEventProcessorTest {
     fun `drops event with the same exception`() {
         val processor = fixture.getSut()
         val event = SentryEvent(RuntimeException())
-        processor.process(event, null)
+        processor.process(event, Hint())
 
-        val result = processor.process(event, null)
+        val result = processor.process(event, Hint())
         assertNull(result)
     }
 
@@ -46,9 +49,12 @@ class DuplicateEventDetectionEventProcessorTest {
     fun `drops event with mechanism exception having an exception that has already been processed`() {
         val processor = fixture.getSut()
         val event = SentryEvent(RuntimeException())
-        processor.process(event, null)
+        processor.process(event, Hint())
 
-        val result = processor.process(SentryEvent(ExceptionMechanismException(Mechanism(), event.throwable!!, Thread.currentThread())), null)
+        val result = processor.process(
+            SentryEvent(ExceptionMechanismException(Mechanism(), event.throwable!!, Thread.currentThread())),
+            Hint()
+        )
         assertNull(result)
     }
 
@@ -56,9 +62,12 @@ class DuplicateEventDetectionEventProcessorTest {
     fun `drops event with exception that has already been processed with event with mechanism exception`() {
         val processor = fixture.getSut()
         val sentryEvent = SentryEvent(ExceptionMechanismException(Mechanism(), RuntimeException(), Thread.currentThread()))
-        processor.process(sentryEvent, null)
+        processor.process(sentryEvent, Hint())
 
-        val result = processor.process(SentryEvent((sentryEvent.throwable as ExceptionMechanismException).throwable), null)
+        val result = processor.process(
+            SentryEvent((sentryEvent.throwable as ExceptionMechanismException).throwable),
+            Hint()
+        )
 
         assertNull(result)
     }
@@ -67,9 +76,12 @@ class DuplicateEventDetectionEventProcessorTest {
     fun `drops event with the cause equal to exception in already processed event`() {
         val processor = fixture.getSut()
         val event = SentryEvent(RuntimeException())
-        processor.process(event, null)
+        processor.process(event, Hint())
 
-        val result = processor.process(SentryEvent(RuntimeException(event.throwable)), null)
+        val result = processor.process(
+            SentryEvent(RuntimeException(event.throwable)),
+            Hint()
+        )
 
         assertNull(result)
     }
@@ -78,9 +90,12 @@ class DuplicateEventDetectionEventProcessorTest {
     fun `drops event with any of the causes has been already processed`() {
         val processor = fixture.getSut()
         val event = SentryEvent(RuntimeException())
-        processor.process(event, null)
+        processor.process(event, Hint())
 
-        val result = processor.process(SentryEvent(RuntimeException(RuntimeException(event.throwable))), null)
+        val result = processor.process(
+            SentryEvent(RuntimeException(RuntimeException(event.throwable))),
+            Hint()
+        )
 
         assertNull(result)
     }
@@ -89,7 +104,7 @@ class DuplicateEventDetectionEventProcessorTest {
     fun `does not deduplicate is deduplication is disabled`() {
         val processor = fixture.getSut(enableDeduplication = false)
         val event = SentryEvent(RuntimeException())
-        assertNotNull(processor.process(event, null))
-        assertNotNull(processor.process(event, null))
+        assertNotNull(processor.process(event, Hint()))
+        assertNotNull(processor.process(event, Hint()))
     }
 }