+++ sentry/src/test/java/io/sentry/HubTest.kt
@@ -14,7 +14,6 @@ import com.nhaarman.mockitokotlin2.times
 import com.nhaarman.mockitokotlin2.verify
 import com.nhaarman.mockitokotlin2.verifyNoMoreInteractions
 import com.nhaarman.mockitokotlin2.whenever
-import io.sentry.TypeCheckHint.SENTRY_TYPE_CHECK_HINT
 import io.sentry.cache.EnvelopeCache
 import io.sentry.clientreport.ClientReportTestHelper.Companion.assertClientReport
 import io.sentry.clientreport.DiscardReason
@@ -266,7 +265,7 @@ class HubTest {
         sut.close()
 
         sut.captureEvent(SentryEvent())
-        verify(mockClient, never()).captureEvent(any(), any<Map<String, Any?>>())
+        verify(mockClient, never()).captureEvent(any(), any<Hint>())
     }
 
     @Test
@@ -274,9 +273,9 @@ class HubTest {
         val (sut, mockClient) = getEnabledHub()
 
         val event = SentryEvent()
-        val hintsMap = mutableMapOf<String, Any>(SENTRY_TYPE_CHECK_HINT to { })
-        sut.captureEvent(event, hintsMap)
-        verify(mockClient).captureEvent(eq(event), any(), eq(hintsMap))
+        val hints = HintUtils.createWithTypeCheckHint({})
+        sut.captureEvent(event, hints)
+        verify(mockClient).captureEvent(eq(event), any(), eq(hints))
     }
 
     @Test
@@ -284,11 +283,11 @@ class HubTest {
         val (sut, mockClient) = getEnabledHub()
         whenever(mockClient.captureEvent(any(), any(), anyOrNull())).thenReturn(SentryId(UUID.randomUUID()))
         val event = SentryEvent()
-        val hintsMap = mutableMapOf<String, Any>(SENTRY_TYPE_CHECK_HINT to { })
-        sut.captureEvent(event, hintsMap)
+        val hints = HintUtils.createWithTypeCheckHint({})
+        sut.captureEvent(event, hints)
         val lastEventId = sut.lastEventId
         sut.close()
-        sut.captureEvent(event, hintsMap)
+        sut.captureEvent(event, hints)
         assertEquals(lastEventId, sut.lastEventId)
     }
 
@@ -297,9 +296,9 @@ class HubTest {
         val (sut, mockClient) = getEnabledHub()
 
         val event = SentryEvent()
-        val hintsMap = mutableMapOf<String, Any>(SENTRY_TYPE_CHECK_HINT to { })
-        sut.captureEvent(event, hintsMap)
-        verify(mockClient).captureEvent(eq(event), any(), eq(hintsMap))
+        val hints = HintUtils.createWithTypeCheckHint({})
+        sut.captureEvent(event, hints)
+        verify(mockClient).captureEvent(eq(event), any(), eq(hints))
         verify(mockClient, never()).captureSession(any(), any())
     }
 
@@ -308,9 +307,9 @@ class HubTest {
         val (sut, mockClient) = getEnabledHub()
 
         val event = SentryEvent()
-        val hintsMap = mutableMapOf<String, Any>(SENTRY_TYPE_CHECK_HINT to { })
-        sut.captureEvent(event, hintsMap)
-        verify(mockClient).captureEvent(eq(event), any(), eq(hintsMap))
+        val hints = HintUtils.createWithTypeCheckHint({})
+        sut.captureEvent(event, hints)
+        verify(mockClient).captureEvent(eq(event), any(), eq(hints))
         verify(mockClient, never()).captureSession(any(), any())
     }
 
@@ -324,10 +323,10 @@ class HubTest {
 
         val event = SentryEvent(exception)
 
-        val hintsMap = mutableMapOf<String, Any>(SENTRY_TYPE_CHECK_HINT to { })
-        sut.captureEvent(event, hintsMap)
+        val hints = HintUtils.createWithTypeCheckHint({})
+        sut.captureEvent(event, hints)
         assertEquals(span.spanContext, event.contexts.trace)
-        verify(mockClient).captureEvent(eq(event), any(), eq(hintsMap))
+        verify(mockClient).captureEvent(eq(event), any(), eq(hints))
     }
 
     @Test
@@ -340,10 +339,10 @@ class HubTest {
 
         val event = SentryEvent(RuntimeException(rootCause))
 
-        val hintsMap = mutableMapOf<String, Any>(SENTRY_TYPE_CHECK_HINT to { })
-        sut.captureEvent(event, hintsMap)
+        val hints = HintUtils.createWithTypeCheckHint({})
+        sut.captureEvent(event, hints)
         assertEquals(span.spanContext, event.contexts.trace)
-        verify(mockClient).captureEvent(eq(event), any(), eq(hintsMap))
+        verify(mockClient).captureEvent(eq(event), any(), eq(hints))
     }
 
     @Test
@@ -357,10 +356,10 @@ class HubTest {
 
         val event = SentryEvent(RuntimeException(exceptionAssignedToSpan))
 
-        val hintsMap = mutableMapOf<String, Any>(SENTRY_TYPE_CHECK_HINT to { })
-        sut.captureEvent(event, hintsMap)
+        val hints = HintUtils.createWithTypeCheckHint({})
+        sut.captureEvent(event, hints)
         assertEquals(span.spanContext, event.contexts.trace)
-        verify(mockClient).captureEvent(eq(event), any(), eq(hintsMap))
+        verify(mockClient).captureEvent(eq(event), any(), eq(hints))
     }
 
     @Test
@@ -375,10 +374,10 @@ class HubTest {
         val originalSpanContext = SpanContext("op")
         event.contexts.trace = originalSpanContext
 
-        val hintsMap = mutableMapOf<String, Any>(SENTRY_TYPE_CHECK_HINT to { })
-        sut.captureEvent(event, hintsMap)
+        val hints = HintUtils.createWithTypeCheckHint({})
+        sut.captureEvent(event, hints)
         assertEquals(originalSpanContext, event.contexts.trace)
-        verify(mockClient).captureEvent(eq(event), any(), eq(hintsMap))
+        verify(mockClient).captureEvent(eq(event), any(), eq(hints))
     }
 
     @Test
@@ -387,10 +386,10 @@ class HubTest {
 
         val event = SentryEvent(RuntimeException())
 
-        val hintsMap = mutableMapOf<String, Any>(SENTRY_TYPE_CHECK_HINT to { })
-        sut.captureEvent(event, hintsMap)
+        val hints = HintUtils.createWithTypeCheckHint({})
+        sut.captureEvent(event, hints)
         assertNull(event.contexts.trace)
-        verify(mockClient).captureEvent(eq(event), any(), eq(hintsMap))
+        verify(mockClient).captureEvent(eq(event), any(), eq(hints))
     }
     //endregion
 
@@ -456,8 +455,8 @@ class HubTest {
     fun `when captureException is called with a valid argument and hint, captureEvent on the client should be called`() {
         val (sut, mockClient) = getEnabledHub()
 
-        val hintsMap = mutableMapOf<String, Any>(SENTRY_TYPE_CHECK_HINT to Object())
-        sut.captureException(Throwable(), hintsMap)
+        val hints = HintUtils.createWithTypeCheckHint({})
+        sut.captureException(Throwable(), hints)
         verify(mockClient).captureEvent(any(), any(), any())
     }
 
@@ -974,7 +973,7 @@ class HubTest {
         sut.bindClient(mockClient)
 
         sut.startSession()
-        verify(mockClient).captureSession(any(), argWhere { HintUtils.getSentrySdkHint(it) is SessionStartHint })
+        verify(mockClient).captureSession(any(), argWhere { HintUtils.hasType(it, SessionStartHint::class.java) })
     }
 
     @Test
@@ -990,8 +989,8 @@ class HubTest {
 
         sut.startSession()
         sut.startSession()
-        verify(mockClient).captureSession(any(), argWhere { HintUtils.getSentrySdkHint(it) is SessionEndHint })
-        verify(mockClient, times(2)).captureSession(any(), argWhere { HintUtils.getSentrySdkHint(it) is SessionStartHint })
+        verify(mockClient).captureSession(any(), argWhere { HintUtils.hasType(it, SessionEndHint::class.java) })
+        verify(mockClient, times(2)).captureSession(any(), argWhere { HintUtils.hasType(it, SessionStartHint::class.java) })
     }
     //endregion
 
@@ -1040,8 +1039,8 @@ class HubTest {
 
         sut.startSession()
         sut.endSession()
-        verify(mockClient).captureSession(any(), argWhere { HintUtils.getSentrySdkHint(it) is SessionStartHint })
-        verify(mockClient).captureSession(any(), argWhere { HintUtils.getSentrySdkHint(it) is SessionEndHint })
+        verify(mockClient).captureSession(any(), argWhere { HintUtils.hasType(it, SessionStartHint::class.java) })
+        verify(mockClient).captureSession(any(), argWhere { HintUtils.hasType(it, SessionEndHint::class.java) })
     }
 
     @Test