+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixes
 
 - No longer close OutputStream that is passed into JsonSerializer (#2029)
+- Fix setting context tags on events captured by Spring (#2060)
 
 ### Features
 
@@ -12,6 +13,8 @@
 - Allow setting native Android SDK name during build ([#2035](https://github.com/getsentry/sentry-java/pull/2035))
 - Include application permissions in Android events ([#2018](https://github.com/getsentry/sentry-java/pull/2018))
 - Automatically create transactions for UI events ([#1975](https://github.com/getsentry/sentry-java/pull/1975))
+- Hints are now used via a Hint object and passed into beforeSend and EventProcessor as @NotNull Hint object ([#2045](https://github.com/getsentry/sentry-java/pull/2045))
+- Attachments can be manipulated via hint ([#2046](https://github.com/getsentry/sentry-java/pull/2046))
 - Calling Sentry.init and specifying contextTags now has an effect on the jul SentryAppender ([#2057](https://github.com/getsentry/sentry-java/pull/2057))
 
 ### Fixes